made some changed in navigation process
