+++ app/src/main/java/co/spacece/neurosign/Navigation.kt
@@ -1,66 +1,30 @@
-package co.spacece.neurosign.Home
+package co.spacece.neurosign
 
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Call
-import androidx.compose.material.icons.filled.Info
-import androidx.compose.material.icons.filled.MoreVert
-import androidx.compose.material3.Icon
-import androidx.compose.material3.NavigationBar
-import androidx.compose.material3.NavigationBarItem
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
-import co.spacece.neurosign.AppTopBar
-import co.spacece.neurosign.Home.Products.NavigationItem
-import co.spacece.neurosign.Home.Products.ProductsScreen
+import co.spacece.neurosign.hobbyPredictor.ui.HobbyPredictorScreen
+import co.spacece.neurosign.hobbyPredictor.ui.ScreenB
+import co.spacece.neurosign.home.products.ProductsScreen
+import co.spacece.neurosign.home.products.ScreenA
 
 
 @Composable
-fun ProductApp() {
+fun Navigation() {
     val navController = rememberNavController()
-
-    Scaffold(
-        topBar = { AppTopBar() },
-        bottomBar = { BottomNavigation() }
-    ) { paddingValues ->
         NavHost(
             navController = navController,
-            startDestination = "products",
-            modifier = Modifier.padding(paddingValues)
+            startDestination = ScreenA,
+            modifier = Modifier.padding()
         ) {
-            composable("products") {
+            composable<ScreenA> {
                 ProductsScreen()
             }
+            composable<ScreenB>{
+                HobbyPredictorScreen()
+            }
         }
-    }
-}
-
-@Composable
-fun BottomNavigation() {
-    NavigationBar {
-        val items = listOf(
-            NavigationItem("About Us", Icons.Filled.Info),
-            NavigationItem("Our Products", Icons.Filled.MoreVert),
-            NavigationItem("Contact Us", Icons.Filled.Call)
-        )
-
-        items.forEach { item ->
-            NavigationBarItem(
-                icon = {
-                    Icon(
-                        imageVector = item.icon, // Use imageVector instead of painter
-                        contentDescription = item.label
-                    )
-                },
-                label = { Text(item.label) },
-                selected = item.label == "Our Products",
-                onClick = { /* TODO: Handle navigation */ }
-            )
-        }
-    }
 }
