+++ app/src/main/java/co/spacece/neurosign/hobbyPredictor/ui/HobbyPredictorScreen.kt
@@ -4,12 +4,19 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import co.spacece.neurosign.AppTopBar
+import kotlinx.serialization.Serializable
 
 
 @Composable
 fun HobbyPredictorScreen(){
-    Scaffold(modifier = Modifier.fillMaxSize()){paddingValues ->
+    Scaffold(modifier = Modifier.fillMaxSize(),
+        topBar = { AppTopBar(title = "Hobby Predictor") }
+    ){paddingValues ->
         HobbyPredictorScreenView()
     }
 
 }
+
+@Serializable
+object ScreenB
