+++ app/src/main/java/co/spacece/neurosign/home/products/ProductCard.kt
@@ -1,4 +1,4 @@
-package co.spacece.neurosign.Home.Products
+package co.spacece.neurosign.home.products
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Arrangement
